add lcsc_territory()
The function generates theoretical territories or catchment areas based on a cost surface for one or more locations. The catchment areas are calculated on the basis of a certain period of time, which, starting from a central location, the movement can take place in all directions depending on the costs.